text2ImageCLI/requirements.txt
requests
pillow
opencv-python
numpy
ollama
pyyaml
rich